Moved Featured resource model to newsroom.types
